+++ CHANGELOG.md
@@ -3,6 +3,11 @@ __In descending date order first__ (most recent date and time first).
 Timestamp Format: __day month HOURS:MINUTES (military time) USER__  
 Example: 10 mar 21:13 SAJID\
 
+## 1 APRIL Jenny
+* Removed Referral inner class from PatientModel and changed referrals to strings instead, for the sake of sanity. 
+* Added a referral section to PatientView and an upload referral option. Upload referral opens up a JFileChooser and the name of the selected file will be added to the list of referrals (String) in PatientModel. This button click event is implemented in PatientController.
+* Currently, the view isn't updated when a file is selected. 
+
 ## 31 MARCH Jenny
 # 21:51
 * Added a Referral inner class to PatientModel. Code is inspired by Eberly's code on Treap in 331. Patients have an ArrayList of Referrals.